refactor: smonta so_server_core.py in 13 moduli dedicati (_inventory, _radar, _topic, ...)
- so_server_core.py passa da 1679 righe monolitiche a 15 righe di facade
- 13 nuovi moduli mcp/_*.py, ognuno con una responsabilità singola
- _artifact.py: path constants, env, artifact resolution, GCS download, helpers condivisi
- _inventory.py, _signals.py, _radar.py, _registry.py, _find_url.py, _probe.py
- _ckan.py, _sparql.py, _html.py, _sdmx.py, _topic.py, _recommend.py
- _inventory.py assorbe _source_radar_context (prima in so_server_core)
- _radar.py assorbe radar_delta (deprecato ma mantenuto per compat)
- _topic.py è pura: zero dipendenze, solo re + unicodedata + typing

Test aggiornati: monkeypatch punti ai moduli corretti (_artifact.*, _probe.*, _html.*)
Inalterati: so_server.py, contratti API, output, interfaccia pubblica
215/215 test passanti, 17 tool MCP registrati
